Trait leptos::ev::EventDescriptor

pub trait EventDescriptor: Clone {
    type EventType: FromWasmAbi;

    const BUBBLES: bool;

    // Required methods
    fn name(&self) -> Cow<'static, str>;
    fn event_delegation_key(&self) -> Cow<'static, str>;

    // Provided method
    fn options(&self) -> &Option<AddEventListenerOptions> { ... }
}

A trait for converting types into web_sys events.

Required Associated Types

type EventType: FromWasmAbi

The web_sys event type, such as web_sys::MouseEvent.

Required Associated Constants

const BUBBLES: bool

Indicates if this event bubbles. For example, click bubbles, but focus does not.

If this is true, then the event will be delegated globally, otherwise, event listeners will be directly attached to the element.

Required Methods

fn name(&self) -> Cow<'static, str>

The name of the event, such as click or mouseover.

fn event_delegation_key(&self) -> Cow<'static, str>

The key used for event delegation.

Provided Methods

fn options(&self) -> &Option<AddEventListenerOptions>

Return the options for this type. This is only used when you create a Custom event handler.