Function leptos::Await

pub fn Await<T, Fut, FF, VF, V>(
    cx: Scope,
    props: AwaitProps<T, Fut, FF, VF, V>
) -> impl IntoViewwhere
    Fut: Future<Output = T> + 'static,
    FF: Fn(Scope) -> Fut + 'static,
    V: IntoView,
    VF: Fn(Scope, &T) -> V + 'static,
    T: Serializable + 'static,

Allows you to inline the data loading for an async block or server function directly into your view. This is the equivalent of combining a create_resource that only loads once (i.e., with a source signal || ()) with a Suspense with no fallback.

async fn fetch_monkeys(monkey: i32) -> i32 {
    // do some expensive work
    3
}
view! { cx,
    <Await
        future=|cx| fetch_monkeys(3)
        view=|cx, data| {
            view! { cx, <p>{*data} " little monkeys, jumping on the bed."</p> }
        }
    />
}

Required Props

cx: Scope

future: [FF]

A function that takes a Scope and returns the Future that will the component will .await before rendering.

view: [VF]

A function that takes a Scope and a reference to the resolved data from the future renders a view.

Optional Props

blocking: bool

If true, the component will use create_blocking_resource, preventing the HTML stream from returning anything before future has resolved.