Trait leptos::html::ElementDescriptor

pub trait ElementDescriptor: ElementDescriptorBounds {
    // Required methods
    fn name(&self) -> Cow<'static, str>;
    fn hydration_id(&self) -> &HydrationKey;

    // Provided method
    fn is_void(&self) -> bool { ... }
}

Trait which allows creating an element tag.

Required Methods

fn name(&self) -> Cow<'static, str>

The name of the element, i.e., div, p, custom-element.

fn hydration_id(&self) -> &HydrationKey

A unique id that should be generated for each new instance of this element, and be consistent for both SSR and CSR.

Provided Methods

fn is_void(&self) -> bool

Determines if the tag is void, i.e., <input> and <br>.