Enum leptos::oco::Oco

pub enum Oco<'a, T>where
    T: ToOwned + 'a + ?Sized,{
    Borrowed(&'a T),
    Counted(Rc<T, Global>),
    Owned(<T as ToOwned>::Owned),
}

“Owned Clones Once” - a smart pointer that can be either a reference, an owned value, or a reference counted pointer. This is useful for storing immutable values, such as strings, in a way that is cheap to clone and pass around.

The Clone implementation is amortized O(1). Cloning the Oco::Borrowed variant simply copies the references (O(1)). Cloning the Oco::Counted variant increments a reference count (O(1)). Cloning the Oco::Owned variant upgrades it to Oco::Counted, which requires an O(n) clone of the data, but all subsequent clones will be O(1).

Variants

Borrowed(&'a T)

A static reference to a value.

Counted(Rc<T, Global>)

A reference counted pointer to a value.

Owned(<T as ToOwned>::Owned)

An owned value.

Implementations

impl<'a, T> Oco<'a, T>where T: ToOwned + ?Sized,

pub fn into_owned(self) -> <T as ToOwned>::Owned

Converts the value into an owned value.

pub const fn is_borrowed(&self) -> bool

Checks if the value is Oco::Borrowed.

Examples

assert!(Oco::<str>::Borrowed("Hello").is_borrowed());
assert!(!Oco::<str>::Counted(Rc::from("Hello")).is_borrowed());
assert!(!Oco::<str>::Owned("Hello".to_string()).is_borrowed());

pub const fn is_counted(&self) -> bool

Checks if the value is Oco::Counted.

Examples

assert!(Oco::<str>::Counted(Rc::from("Hello")).is_counted());
assert!(!Oco::<str>::Borrowed("Hello").is_counted());
assert!(!Oco::<str>::Owned("Hello".to_string()).is_counted());

pub const fn is_owned(&self) -> bool

Checks if the value is Oco::Owned.

Examples

assert!(Oco::<str>::Owned("Hello".to_string()).is_owned());
assert!(!Oco::<str>::Borrowed("Hello").is_owned());
assert!(!Oco::<str>::Counted(Rc::from("Hello")).is_owned());

impl Oco<'_, str>

pub fn as_str(&self) -> &str

Returns a &str slice of this Oco.

Examples

let oco = Oco::<str>::Borrowed("Hello");
let s: &str = oco.as_str();
assert_eq!(s, "Hello");

impl Oco<'_, CStr>

pub fn as_c_str(&self) -> &CStr

Returns a &CStr slice of this Oco.

Examples

use std::ffi::CStr;

let oco =
    Oco::<CStr>::Borrowed(CStr::from_bytes_with_nul(b"Hello\0").unwrap());
let s: &CStr = oco.as_c_str();
assert_eq!(s, CStr::from_bytes_with_nul(b"Hello\0").unwrap());

impl Oco<'_, OsStr>

pub fn as_os_str(&self) -> &OsStr

Returns a &OsStr slice of this Oco.

Examples

use std::ffi::OsStr;

let oco = Oco::<OsStr>::Borrowed(OsStr::new("Hello"));
let s: &OsStr = oco.as_os_str();
assert_eq!(s, OsStr::new("Hello"));

impl Oco<'_, Path>

pub fn as_path(&self) -> &Path

Returns a &Path slice of this Oco.

Examples

use std::path::Path;

let oco = Oco::<Path>::Borrowed(Path::new("Hello"));
let s: &Path = oco.as_path();
assert_eq!(s, Path::new("Hello"));

impl<T> Oco<'_, [T]>where [T]: ToOwned,

pub fn as_slice(&self) -> &[T]

Returns a &[T] slice of this Oco.

Examples

let oco = Oco::<[u8]>::Borrowed(b"Hello");
let s: &[u8] = oco.as_slice();
assert_eq!(s, b"Hello");

impl<'a, T> Oco<'a, T>where T: ToOwned + 'a + ?Sized, Rc<T, Global>: for<'b> From<&'b T>,

pub fn clone_inplace(&mut self) -> Oco<'a, T>

Clones the value with inplace conversion into Oco::Counted if it was previously Oco::Owned.

Examples

let mut oco1 = Oco::<str>::Owned("Hello".to_string());
let oco2 = oco1.clone_inplace();
assert_eq!(oco1, oco2);
assert!(oco1.is_counted());
assert!(oco2.is_counted());