Module leptos::signal_prelude

This prelude imports all signal types as well as all signal traits needed to use those types.

Structs

Memo

An efficient derived reactive value based on other reactive values.

ReadSignal

The getter for a reactive signal.

RwSignal

A signal that combines the getter and setter into one value, rather than separating them into a ReadSignal and a WriteSignal. You may prefer this its style, or it may be easier to pass around in a context or as a function argument.

Signal

A wrapper for any kind of readable reactive signal: a ReadSignal, Memo, RwSignal, or derived signal closure.

SignalSetter

A wrapper for any kind of settable reactive signal: a WriteSignal, RwSignal, or closure that receives a value and sets a signal depending on it.

WriteSignal

The setter for a reactive signal.

Enums

MaybeSignal

A wrapper for a value that is either T or Signal<T>.

Traits

IntoSignal

Helper trait for converting Fn() -> T closures into Signal<T>.

IntoSignalSetter

Helper trait for converting Fn(T) into SignalSetter<T>.

SignalDispose

This trait allows disposing a signal before its Scope has been disposed.

SignalGet

This trait allows getting an owned value of the signals inner type.

SignalGetUntracked

Trait implemented for all signal types which you can get a value from, such as ReadSignal, Memo, etc., which allows getting the inner value without subscribing to the current scope.

SignalSet

This trait allows setting the value of a signal.

SignalSetUntracked

Trait implemented for all signal types which you can set the inner value, such as WriteSignal and RwSignal, which allows setting the inner value without causing effects which depend on the signal from being run.

SignalStream

This trait allows converting a signal into a async Stream.

SignalUpdate

This trait allows updating the inner value of a signal.

SignalUpdateUntracked

This trait allows updating the signals value without causing dependant effects to run.

SignalWith

This trait allows obtaining an immutable reference to the signal’s inner type.

SignalWithUntracked

This trait allows getting a reference to the signals inner value without creating a dependency on the signal.

Functions

create_many_signals

Works exactly as create_signal, but creates multiple signals at once.

create_many_signals_mapped

Works exactly as create_many_signals, but applies the map function to each signal pair.

create_memo

Creates an efficient derived reactive value based on other reactive values.

create_rw_signal

Creates a reactive signal with the getter and setter unified in one value. You may prefer this style, or it may be easier to pass around in a context or as a function argument.

create_selector

Creates a conditional signal that only notifies subscribers when a change in the source signal’s value changes whether it is equal to the key value (as determined by PartialEq.)

create_selector_with_fn

Creates a conditional signal that only notifies subscribers when a change in the source signal’s value changes whether the given function is true.

create_signal

Creates a signal, the basic reactive primitive.

create_signal_from_stream

Creates a signal that always contains the most recent value emitted by a Stream. If the stream has not yet emitted a value since the signal was created, the signal’s value will be None.