Struct leptos::signal_prelude::MaybeProp

pub struct MaybeProp<T>(/* private fields */)
where
    T: 'static;

A wrapping type for an optional component prop, which can either be a signal or a non-reactive value, and which may or may not have a value. In other words, this is an Option<MaybeSignal<Option<T>>> that automatically flattens its getters.

This creates an extremely flexible type for component libraries, etc.

Core Trait Implementations

.get() (or calling the signal as a function) clones the current value of the signal. If you call it within an effect, it will cause that effect to subscribe to the signal, and to re-run whenever the value of the signal changes.

.get_untracked() clones the value of the signal without reactively tracking it.

.with() allows you to reactively access the signal’s value without cloning by applying a callback function.

.with_untracked() allows you to access the signal’s value without reactively tracking it.

.to_stream() converts the signal to an async stream of values.

Examples

let (count, set_count) = create_signal(Some(2));
let double = |n| n * 2;
let double_count = MaybeProp::derive(move || count.get().map(double));
let memoized_double_count = create_memo(move |_| count.get().map(double));
let static_value = 5;

// this function takes either a reactive or non-reactive value
fn above_3(arg: &MaybeProp<i32>) -> bool {
    // ✅ calling the signal clones and returns the value
    //    it is a shorthand for arg.get()q
    arg.get().map(|arg| arg > 3).unwrap_or(false)
}

assert_eq!(above_3(&None::<i32>.into()), false);
assert_eq!(above_3(&static_value.into()), true);
assert_eq!(above_3(&count.into()), false);
assert_eq!(above_3(&double_count), true);
assert_eq!(above_3(&memoized_double_count.into()), true);

Implementations

impl<T> MaybeProp<T>

Examples

let (name, set_name) = create_signal("Alice".to_string());
let (maybe_name, set_maybe_name) = create_signal(None);
let name_upper =
    MaybeProp::derive(move || Some(name.with(|n| n.to_uppercase())));
let memoized_lower = create_memo(move |_| name.with(|n| n.to_lowercase()));
let static_value: MaybeProp<String> = "Bob".to_string().into();

// this function takes any kind of wrapped signal
fn current_len_inefficient(arg: &MaybeProp<String>) -> usize {
    // ❌ unnecessarily clones the string
    arg.get().map(|n| n.len()).unwrap_or(0)
}

fn current_len(arg: &MaybeProp<String>) -> usize {
    // ✅ gets the length without cloning the `String`
    arg.with(|value| value.len()).unwrap_or(0)
}

assert_eq!(current_len(&None::<String>.into()), 0);
assert_eq!(current_len(&maybe_name.into()), 0);
assert_eq!(current_len(&name_upper), 5);
assert_eq!(current_len(&memoized_lower.into()), 5);
assert_eq!(current_len(&static_value), 3);

// Normal signals/memos return T
assert_eq!(name.get(), "Alice".to_string());
assert_eq!(memoized_lower.get(), "alice".to_string());

// MaybeProp::get() returns Option<T>
assert_eq!(name_upper.get(), Some("ALICE".to_string()));
assert_eq!(static_value.get(), Some("Bob".to_string()));

pub fn with<O>(&self, f: impl FnOnce(&T) -> O) -> Option<O>

Applies a function to the current value, returning the result.

pub fn try_with<O>(&self, f: impl FnOnce(&T) -> O) -> Option<O>

Applies a function to the current value, returning the result. Returns None if the value has already been disposed.

pub fn with_untracked<O>(&self, f: impl FnOnce(&T) -> O) -> Option<O>

Applies a function to the current value, returning the result, without causing the current reactive scope to track changes.

pub fn try_with_untracked<O>(&self, f: impl FnOnce(&T) -> O) -> Option<O>

Applies a function to the current value, returning the result, without causing the current reactive scope to track changes. Returns None if the value has already been disposed.

impl<T> MaybeProp<T>where T: 'static,

pub fn derive(derived_signal: impl Fn() -> Option<T> + 'static) -> MaybeProp<T>

Wraps a derived signal, i.e., any computation that accesses one or more reactive signals.

let (count, set_count) = create_signal(2);
let double_count = MaybeProp::derive(move || Some(count.get() * 2));

// this function takes any kind of wrapped signal
fn above_3(arg: &MaybeProp<i32>) -> bool {
    arg.get().unwrap_or(0) > 3
}

assert_eq!(above_3(&count.into()), false);
assert_eq!(above_3(&double_count.into()), true);
assert_eq!(above_3(&2.into()), false);