Trait leptos::signal_prelude::SignalGet

pub trait SignalGet<T> {
    // Required methods
    fn get(&self) -> T;
    fn try_get(&self) -> Option<T>;
}

This trait allows getting an owned value of the signals inner type.

Required Methods

fn get(&self) -> T

Clones and returns the current value of the signal, and subscribes the running effect to this signal.

Panics

Panics if you try to access a signal that was created in a Scope that has been disposed.

fn try_get(&self) -> Option<T>

Clones and returns the signal value, returning Some if the signal is still alive, and None otherwise.