Trait leptos::signal_prelude::SignalSet

pub trait SignalSet<T> {
    // Required methods
    fn set(&self, new_value: T);
    fn try_set(&self, new_value: T) -> Option<T>;
}

This trait allows setting the value of a signal.

Required Methods

fn set(&self, new_value: T)

Sets the signal’s value and notifies subscribers.

Note: set() does not auto-memoize, i.e., it will notify subscribers even if the value has not actually changed.

fn try_set(&self, new_value: T) -> Option<T>

Sets the signal’s value and notifies subscribers. Returns None if the signal is still valid, [Some(T)] otherwise.

Note: set() does not auto-memoize, i.e., it will notify subscribers even if the value has not actually changed.