Module leptos::ssr

Utilities for server-side rendering HTML.

Functions

render_to_stream

Renders a function to a stream of HTML strings.

render_to_stream_in_order

Renders an in-order HTML stream, pausing at <Suspense/> components. The stream contains, in order:

render_to_stream_in_order_with_prefix

Renders an in-order HTML stream, pausing at <Suspense/> components. The stream contains, in order:

render_to_stream_in_order_with_prefix_undisposed_with_context

Renders an in-order HTML stream, pausing at <Suspense/> components. The stream contains, in order:

render_to_stream_with_prefix

Renders a function to a stream of HTML strings. After the view runs, the prefix will run with the same scope. This can be used to generate additional HTML that has access to the same reactive graph.

render_to_stream_with_prefix_undisposed

Renders a function to a stream of HTML strings and returns the RuntimeId that was created, so it can be disposed when appropriate. After the view runs, the prefix will run with the same scope. This can be used to generate additional HTML that has access to the same reactive graph.

render_to_stream_with_prefix_undisposed_with_context

Renders a function to a stream of HTML strings and returns the RuntimeId that was created, so they can be disposed when appropriate. After the view runs, the prefix will run with the same scope. This can be used to generate additional HTML that has access to the same reactive graph.

render_to_stream_with_prefix_undisposed_with_context_and_block_replacement

Renders a function to a stream of HTML strings and returns the RuntimeId that was created, so they can be disposed when appropriate. After the view runs, the prefix will run with the same scope. This can be used to generate additional HTML that has access to the same reactive graph.

render_to_string

Renders the given function to a static HTML string.

render_to_string_async

Renders a view to HTML, waiting to return until all async Resources loaded in <Suspense/> elements have finished loading.