Struct leptos::Trigger

pub struct Trigger { /* private fields */ }

Reactive Trigger, notifies reactive code to rerun.

See create_trigger for more.

Implementations

impl Trigger

pub fn notify(&self)

Notifies any reactive code where this trigger is tracked to rerun.

pub fn try_notify(&self) -> bool

Attempts to notify any reactive code where this trigger is tracked to rerun.

Returns None if the runtime has been disposed.

pub fn track(&self)

Subscribes the running effect to this trigger.

pub fn try_track(&self) -> bool

Attempts to subscribe the running effect to this trigger, returning None if the runtime has been disposed.