Trait leptos::SignalStream

pub trait SignalStream<T> {
    // Required method
    fn to_stream(
        &self,
        cx: Scope
    ) -> Pin<Box<dyn Stream<Item = T> + 'static, Global>>;
}

This trait allows converting a signal into a async Stream.

Required Methods

fn to_stream( &self, cx: Scope ) -> Pin<Box<dyn Stream<Item = T> + 'static, Global>>

Generates a Stream that emits the new value of the signal whenever it changes.

Panics

Panics if you try to access a signal that was created in a Scope that has been disposed.