Trait leptos::SignalWith

pub trait SignalWith {
    type Value;

    // Required methods
    fn with<O>(&self, f: impl FnOnce(&Self::Value) -> O) -> O;
    fn try_with<O>(&self, f: impl FnOnce(&Self::Value) -> O) -> Option<O>;

    // Provided method
    fn track(&self) { ... }
}

This trait allows obtaining an immutable reference to the signal’s inner type.

Required Associated Types

type Value

The value held by the signal.

Required Methods

fn with<O>(&self, f: impl FnOnce(&Self::Value) -> O) -> O

Applies a function to the current value of the signal, and subscribes the running effect to this signal.

Panics

Panics if you try to access a signal that is owned by a reactive node that has been disposed.

fn try_with<O>(&self, f: impl FnOnce(&Self::Value) -> O) -> Option<O>

Applies a function to the current value of the signal, and subscribes the running effect to this signal. Returns Some if the signal is valid and the function ran, otherwise returns None.

Provided Methods

fn track(&self)

Subscribes to this signal in the current reactive scope without doing anything with its value.