Trait leptos::signal_prelude::SignalGet

pub trait SignalGet {
    type Value;

    // Required methods
    fn get(&self) -> Self::Value;
    fn try_get(&self) -> Option<Self::Value>;
}

This trait allows getting an owned value of the signals inner type.

Required Associated Types

type Value

The value held by the signal.

Required Methods

fn get(&self) -> Self::Value

Clones and returns the current value of the signal, and subscribes the running effect to this signal.

Panics

Panics if you try to access a signal that is owned by a reactive node that has been disposed.

fn try_get(&self) -> Option<Self::Value>

Clones and returns the signal value, returning Some if the signal is still alive, and None otherwise.